CUDA field calculations now done in single precision.

+++ include/globals.h
@@ -46,7 +46,7 @@ namespace globals
   GLOBAL Array2D<double> h;
   GLOBAL Array2D<double> w;
 
-  GLOBAL SparseMatrix<double> Jij;
+  GLOBAL SparseMatrix<float> Jij;
 
   GLOBAL Array<double> alpha;
   GLOBAL Array<double> mus;
